Show actual sensor values in controller properties dialog
Group three JoyAxis objects together in JoySensor and use the axis
objects to perform the GUI update logic.

Enable sensor events in input daemon, handle them in all input
passes and finally activate them at the end of the last pass.
The event data is converted to int by JoySensor and passed to the
pseudo axes for further processing.

Query sensors during SetJoystick initialization and assign their axes to
the new JoySensor objects. Add sensor axes to a new QHash in SetJoystick to
avoid duplicate IDs and gaps in the ID range which cause various problems in
existing logic.