Add sensor buttons to JoyTabWidget
If a controller has sensors, add configuration buttons similar to a
control stick to the main view. Add required new classes and add
boilerplate. The layout is based on a isometric 3D view with the regular
XY axes and a diagonal Z axis.

The new buttons have no functionality yet.