Implement gyroscope calibration
Gyroscope calibration is mandatory to get useful information from it.
Add a new calibration dialog for sensors because the existing one
appears to be buggy and the workflow for sensors is different.

Apply calibration to all sets because it does not make sense to have
different calibration values for different mapping sets. Hence store the
sensor calibration data outside of the "sets" XML tree.