Consistently use radians and degree in JoySensor
Use radians for all member variables since radians are required for the
calculations every tick. Use degree for all getters, setters and
constants as this is used by the user interface.
Use qFuzzyCompare when comparing back and forth converted floating point
numbers because they will most likely not equal to the original due to
rounding errors.