Implement accelerometer shock detection
Read the sensor data rate which is necessary for the calculations in
case of recent SDL. If no data rate is available assume a defualt of
5 ms to avoid division by zero. Using some type of timer does not work
because none of SDL ticks, QElapsedTimer and std::chrono::high_resolution_timer
had enough resolution to be non-zero between two events.

Add a first order lag filter which processes the raw sensor data. The
detailed detection process is explained in the source code.